snowflake_transfer/requirements.txt
pandas
python-dotenv
snowflake-connector-python
sqlalchemy 
snowflake-sqlalchemy
snowflake-connector-python[pandas]
pathlib
snowflake